feature/35070_event_contracts / 2
Fix variables format

+++ contracts/Main.sol
@@ -17,8 +17,8 @@ contract Main {
 
     event NewEvent(string eventName, uint256 indexed createdTimestamp, address indexed eventAddress, address indexed eventCreator);
 
-    function newEvent(string name, uint256 deposit, bytes2 locale, bytes32 category, string[] tags, string description,
-        uint operatorId, uint64 startDate, uint64 endDate, string sourceUrl, string[] possibleResults
+    function newEvent(string name, uint deposit, bytes2 locale, bytes32 category, string[] tags, string description,
+        uint operatorId, uint startDate, uint endDate, string sourceUrl, string[] possibleResults
     ) returns (address) {
 
         Event lastEvent = new Event(msg.sender, name, deposit, locale, category, tags, description, startDate, endDate,
